+++ colossalai/zero/sharded_param/__init__.py
@@ -1,4 +1,11 @@
 from colossalai.zero.sharded_param.sharded_tensor import ShardedTensor
 from colossalai.zero.sharded_param.sharded_param import ShardedParamV2
+from colossalai.zero.sharded_param.tensor_utils import (colo_model_data_tensor_move_inline, colo_model_data_move_to_cpu,
+                                                        colo_model_tensor_clone, colo_tensor_mem_usage)
+from colossalai.zero.sharded_param.tensorful_state import TensorState, StatefulTensor
 
-__all__ = ['ShardedTensor', 'ShardedParamV2']
+__all__ = [
+    'ShardedTensor', 'ShardedParamV2'
+    'colo_model_data_tensor_move_inline', 'colo_model_data_move_to_cpu', 'colo_model_tensor_clone',
+    'colo_tensor_mem_usage', 'TensorState', 'StatefulTensor'
+]
